third_party/WebKit/LayoutTests/fast/dom/shadow/shadowroot-host-expected.txt

CONSOLE WARNING: line 14: Calling Element.createShadowRoot() for an element which already hosts a shadow root is deprecated and will be removed in M63, around December 2017. See https://www.chromestatus.com/features/4668884095336448 for more details.
Tests ShadowRoot.host property

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



PASS root.host is host
PASS roots[0].host is host
PASS roots[1].host is host
PASS roots[0].host is host
PASS roots[1].host is host
PASS roots[1].host is host
PASS roots[6].host is host
PASS root.host is host



PASS successfullyParsed is true

TEST COMPLETE
